code/requirements.txt
pydantic>=2.5.0
pyyaml>=6.0.1
typer>=0.12.0
apache-iotdb>=1.3.0
